Removed items overlapping
If player drags item A and drops it over item B, item A will continue going to the mouse's position until it's no longer overlapping with another item or if the player stops pressing left click.  If the player stops pressing left click while overlapping another item, item A  will continue to follow the mouse until it reaches a tile with no item.